# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Traveloop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip by ID
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /trips/{tripId}/share:
    patch:
      operationId: toggleTripShare
      tags: [trips]
      summary: Toggle trip public/private sharing
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripShareToggle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/public/{shareCode}:
    get:
      operationId: getPublicTrip
      tags: [trips]
      summary: Get a publicly shared trip by share code
      parameters:
        - name: shareCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Stops
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Trip:
      type: object
      required: [id, userId, name, startDate, endDate, createdAt, isPublic, stopCount]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        coverPhoto:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        isPublic:
          type: boolean
        shareCode:
          type: ["string", "null"]
        totalBudget:
          type: ["number", "null"]
        stopCount:
          type: integer
        createdAt:
          type: string
    TripDetail:
      type: object
      required: [id, userId, name, startDate, endDate, createdAt, isPublic, stops]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        coverPhoto:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        isPublic:
          type: boolean
        shareCode:
          type: ["string", "null"]
        totalBudget:
          type: ["number", "null"]
        createdAt:
          type: string
        stops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
    TripInput:
      type: object
      required: [name, startDate, endDate]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        coverPhoto:
          type: string
        totalBudget:
          type: number
    TripUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        coverPhoto:
          type: string
        totalBudget:
          type: number
    TripShareToggle:
      type: object
      required: [isPublic]
      properties:
        isPublic:
          type: boolean
...
    Stop:
      type: object
      required: [id, tripId, cityId, cityName, country, startDate, endDate, order]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        cityId:
          type: integer
        cityName:
          type: string
        country:
          type: string
        cityImageUrl:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        order:
          type: integer
        activities:
          type: array
          items:
            $ref: "#/components/schemas/StopActivity"
...
    StopActivity:
      type: object
      required: [id, stopId, name, category]
      properties:
        id:
          type: integer
        stopId:
          type: integer
        activityCatalogId:
          type: ["integer", "null"]
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        duration:
          type: ["number", "null"]
        scheduledTime:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
...